+++ lib/simple_code.dart
@@ -5,71 +5,28 @@ import 'dart:math';
 
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/foundation.dart';
+import 'package:flutter/foundation.dart' as foundation;
 import 'package:flutter/material.dart';
 import 'package:flutter/material.dart' as material;
-import 'package:flutter/foundation.dart' as foundation;
+import 'package:flutter/services.dart';
+import 'package:logger/logger.dart';
+import 'package:simple_animations/simple_animations.dart';
+import 'package:time/time.dart';
+import 'package:url_launcher/url_launcher.dart';
 
-part 'simple_navigator/navigator.dart';
-part 'simple_navigator/route.dart';
-part 'simple_utils.dart';
-part 'simple_widgets/animated_button.dart';
-part 'simple_widgets/animated_tween.dart';
-part 'simple_widgets/expanded_icon.dart';
-part 'simple_widgets/expanded_text.dart';
-
-class SimpleCode {
-  static GlobalKey<NavigatorState> navigatorKey = GlobalKey<NavigatorState>();
-  static _Navigator navigator = _Navigator(navigatorKey);
-  static Size _emuSize = const Size(360, 640);
-  static BuildContext get _context =>
-      navigatorKey?.currentState?.overlay?.context;
-
-  static Size get _screenSize => MediaQuery.of(_context).size;
-
-  // ignore: use_setters_to_change_properties
-  static void changeEmulatorSize(Size size) => _emuSize = size;
-
-  // ignore: use_setters_to_change_properties
-  static void changeNavigatorKey(GlobalKey<NavigatorState> newKey) =>
-      navigatorKey = newKey;
-
-  static void _checkKey() {
-    if (_context == null) {
-      throw 'set navigatorKey: SimpleCode.navigatorKey on MaterialApp';
-    }
-  }
-
-  static double _wsz(double size) {
-    _checkKey();
-    return (size / _emuSize.width) * _screenSize.width;
-  }
-
-  static double _hsz(double size) {
-    _checkKey();
-    return (size / _emuSize.height) * _screenSize.height;
-  }
-
-  static double _sz(double size) {
-    if (size < 0) {
-      final double aux = size * -1;
-      return -min(_wsz(aux), _hsz(aux));
-    }
-    return min(_wsz(size), _hsz(size));
-  }
-
-  // static bool _isVertical(Size size) => size.width <= size.height;
-
-  // static bool _isHorizontal(Size size) => size.height < size.width;
-}
-
-double sz(double size) {
-  return SimpleCode._sz(size);
-}
-
-double wsz(double size) {
-  return SimpleCode._wsz(size);
-}
-
-double hsz(double size) {
-  return SimpleCode._hsz(size);
-}
+part 'src/extensions/bool_extension.dart';
+part 'src/extensions/context_extension.dart';
+part 'src/extensions/iterable_extension.dart';
+part 'src/extensions/num_extension.dart';
+part 'src/extensions/string_extension.dart';
+part 'src/extensions/tween_extension.dart';
+part 'src/simple_code.dart';
+part 'src/simple_navigator/navigator.dart';
+part 'src/simple_navigator/route.dart';
+part 'src/simple_utils.dart';
+part 'src/simple_widgets/animated_button.dart';
+part 'src/simple_widgets/animated_tween.dart';
+part 'src/simple_widgets/animation.dart';
+part 'src/simple_widgets/expanded_icon.dart';
+part 'src/simple_widgets/expanded_text.dart';
+part 'src/simple_widgets/tap_animation.dart';