+++ lib/src/extensions/bool_extension.dart
@@ -1,0 +1,5 @@
+part of simple_code;
+
+extension BoolExtension on bool {
+  bool get toggle => !this;
+}